requirements-prod.txt
# Echoelmusic Production Dependencies
# =====================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Cache & Queue
redis==5.0.1
celery==5.3.4

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# AI/ML
torch==2.1.2
torchvision==0.16.2
transformers==4.36.2
diffusers==0.25.0
accelerate==0.25.0
safetensors==0.4.1
einops==0.7.0

# Video Processing
opencv-python-headless==4.9.0.80
imageio==2.33.1
imageio-ffmpeg==0.4.9

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-prometheus==0.43b0
structlog==24.1.0

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Utilities
numpy==1.26.3
pillow==10.2.0
python-dotenv==1.0.0
tenacity==8.2.3
orjson==3.9.12

# Security
cryptography==41.0.7

# Testing (needed for CI)
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
